Formatting lint error

+++ drivers/led/issi/is31fl3729-simple.h
@@ -23,7 +23,7 @@
 #include "progmem.h"
 
 typedef struct is31_led {
-    uint32_t driver : 2;
+    uint8_t driver : 2;
     uint8_t v;
 } __attribute__((packed)) is31_led;
